requirements.txt
playwright==1.49.0
sqlalchemy
asyncpg
greenlet
python-dotenv
fastapi
uvicorn[standard]
pydantic
redis
celery
psycopg2-binary
numpy
huggingface-hub
torch
transformers
sentence-transformers
requests
httpx

